Add a --no-confirmation option
This might be required for some complex release pipelines where the package is
already guaranteed to be well reviewed and checked before `opam-publish` is
called.

I took care to add some safeguards to avoid copy-paste of ill-advised one-liners
that would result in people submitting unreviewed releases and increasing the
strain on the repository maintainers.