Run opam-ci-check lints before submitting a package
As a part of our work to improve package authoring workflows, we have
extracted some of the functionality from the opam-repository CI code
into a CLI sub-project, opam-ci-check, that can be used locally to run
some of the CI steps, including linting.

This commit enables running the opam-ci-check lints on a package before
submitting it to opam-repository. It also adds a flag to enable running
these lints when submitting to a different repository. The `--force`
argument can be used to submit the package inspite of any lint errors.